CVS-45090 Adjust ModelInstance class to fulfill common interface for both stateless and stateful models (#429)
* Moved inference logic from predict service utils to model instance.
* Moved model management utils (getters, reload methods) from predict service utils to model instance
* ExecutingStreamGuard now holds infer request reference as it's member
