+++ src/BUILD
@@ -29,7 +29,7 @@ cc_library(
         "dl_node.hpp",
         "entry_node.cpp",
         "entry_node.hpp",
-        "executinstreamidguard.hpp",
+        "executingstreamidguard.hpp",
         "exit_node.cpp",
         "exit_node.hpp",
         "filesystem.hpp",
@@ -199,7 +199,6 @@ cc_test(
         "test/pipelinedefinitionstatus_test.cpp",
         "test/predict_validation_test.cpp",
         "test/prediction_service_test.cpp",
-        "test/prediction_service_utils_test.cpp",
         "test/custom_loader_test.cpp",
         "test/rest_parser_row_test.cpp",
         "test/rest_parser_column_test.cpp",
