+++ src/dl_node.hpp
@@ -20,7 +20,7 @@
 #include <string>
 #include <unordered_map>
 
-#include "executinstreamidguard.hpp"
+#include "executingstreamidguard.hpp"
 #include "model_version_policy.hpp"  // for model_version_t typename
 #include "modelinstance.hpp"
 #include "modelinstanceunloadguard.hpp"
