+++ src/executingstreamidguard.hpp
@@ -21,14 +21,17 @@ namespace ovms {
 struct ExecutingStreamIdGuard {
     ExecutingStreamIdGuard(ovms::OVInferRequestsQueue& inferRequestsQueue) :
         inferRequestsQueue_(inferRequestsQueue),
-        id_(inferRequestsQueue_.getIdleStream().get()) {}
+        id_(inferRequestsQueue_.getIdleStream().get()),
+        inferRequest(inferRequestsQueue.getInferRequest(id_)) {}
     ~ExecutingStreamIdGuard() {
         inferRequestsQueue_.returnStream(id_);
     }
     int getId() { return id_; }
+    InferenceEngine::InferRequest& getInferRequest() { return inferRequest; }
 
 private:
     ovms::OVInferRequestsQueue& inferRequestsQueue_;
     const int id_;
+    InferenceEngine::InferRequest& inferRequest;
 };
 }  //  namespace ovms
