+++ src/ovinferrequestsqueue.hpp
@@ -18,6 +18,7 @@
 #include <atomic>
 #include <condition_variable>
 #include <future>
+#include <memory>
 #include <mutex>
 #include <queue>
 #include <thread>
